pyproject.toml
[tool.poetry]
name = "mapf-web-db-migration"
version = "0.1.0"
description = ""
authors = ["Will Holdsworth <83637728+wi11-holdsworth@users.noreply.github.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
regex = "^2024.5.15"
pyinstrument = "^4.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
